Use new components on the Earn page
* used new header version, depending on Earn stage
* removed unused "Coming soon" page
* refactored existing code, extracted smaller components, cleaned up
   styles
* created hardcoded Earn stages to test them easily
