+++ .env.defaults
@@ -12,7 +12,6 @@ HIDE_SWAP_REWARDS=true
 GAS_PRICE_POOLING_FREQUENCY=120
 ETHEREUM_NETWORK=mainnet
 PERSIST_UI_LOCATION=true
-EARN_COMING_SOON=true
 WEBSITE_ORIGIN=https://tally.cash
 USE_MAINNET_FORK=false
 MAINNET_FORK_URL="ws://mainnet-fork.tally.cash:2096/"
