+++ background/features.ts
@@ -5,7 +5,6 @@ export const HIDE_IMPORT_DERIVATION_PATH =
 export const HIDE_CREATE_PHRASE = process.env.HIDE_CREATE_PHRASE === "true"
 export const PERSIST_UI_LOCATION = process.env.PERSIST_UI_LOCATION === "true"
 export const USE_MAINNET_FORK = process.env.USE_MAINNET_FORK === "true"
-export const EARN_COMING_SOON = process.env.EARN_COMING_SOON === "true"
 export const HIDE_SWAP_REWARDS = process.env.HIDE_SWAP_REWARDS === "true"
 export const MULTI_NETWORK = process.env.MULTI_NETWORK === "true"
 export const CUSTOM_GAS_SELECT = process.env.CUSTOM_GAS_SELECT === "true"
