+++ ui/components/Earn/EmptyBowl.tsx
