+++ ui/routes/routes.tsx
@@ -1,5 +1,4 @@
 import React, { ReactElement } from "react"
-import { EARN_COMING_SOON } from "@tallyho/tally-background/features"
 import Wallet from "../pages/Wallet"
 import SignTransaction from "../pages/SignTransaction"
 import SignData from "../pages/SignData"
@@ -14,7 +13,6 @@ import OnboardingInterstitialCreatePhrase from "../pages/Onboarding/OnboardingIn
 import Overview from "../pages/Overview"
 import SingleAsset from "../pages/SingleAsset"
 import Earn from "../pages/Earn"
-import ComingSoon from "../pages/ComingSoon"
 import EarnDeposit from "../pages/EarnDeposit"
 import Menu from "../pages/Menu"
 import Send from "../pages/Send"
@@ -123,7 +121,7 @@ const pageList: PageList[] = [
   },
   {
     path: "/earn",
-    Component: EARN_COMING_SOON ? ComingSoon : Earn,
+    Component: Earn,
     hasTabBar: true,
     hasTopBar: true,
     persistOnClose: true,
